Move unused classes to different packages
