+++ TeamCode/src/main/java/testcode/DistanceTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package testcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
